Target file same name as source file, chunk size is 4096. Line in urls.txt is striped.

+++ slanje.py
@@ -17,16 +17,41 @@ script to save served file localy.
 """
 
 import requests
+import sys
+import os
+import ntpath
 
 
-#url = 'http://192.168.110.68:8000/text.txt'
-url = 'http://localhost:8000/report.txt'
-target_path = 'my_text.txt'
+urls_file = 'urls.txt'
 
-# stream is true, for big files, from here:
-# http://masnun.com/2016/09/18/python-using-the-requests-module-to-download-large-files-efficiently.html
-response = requests.get(url, stream=True)
-handle = open(target_path, "wb")
-for chunk in response.iter_content(chunk_size=512):
-    if chunk:  # filter out keep-alive new chunks
-        handle.write(chunk)
+# # First command line argument is a text file with urls
+# if len(sys.argv) > 1:
+#     text_file = sys.argv[1]
+# else:
+#     sys.exit("Prisilno gašenje")
+
+def path_leaf(path):
+    """
+    :param path: Take path as argument
+    :return: File name or if no file name, return subdirectory
+    """
+    head, tail = ntpath.split(path)
+    return tail or ntpath.basename(head)
+
+if os.path.exists(urls_file):
+    with open(urls_file, 'r') as f:
+        for line in f:
+            try:
+                url = line.strip()
+                # save target file with the same name as source file
+                target_path = path_leaf(url)
+                # stream is true, for big files, from here:
+                # http://masnun.com/2016/09/18/python-using-the-requests-module-to-download-large-files-efficiently.html
+                response = requests.get(url, stream=True)
+                handle = open(target_path, "wb")
+                for chunk in response.iter_content(chunk_size=4096):
+                    if chunk:  # filter out keep-alive new chunks
+                        handle.write(chunk)
+            except IOError as e:
+                print("Could not read file: {}".format(url))
+                pass
